Black Frogs and Force Applications
The black frogs now behave as intended, but due to the transforms not being zeroed out on the parent game objects properly, the black frogs in the coin pusher are all out of place. The Force application to the player when colliding with the frog hitboxes has been amended, and the forces can be changed in the inspector. In the Coin Pusher scene, the Enemies GameObject has been rearranged and its transform has been zeroed out properly.