Input Fixes, Manhole Proxy and Level Updates
The Input bugs have been largely fixed, but some oddities exist with transforms and ground checks that will need collaboration with the level designers to solve. A proxy of the manhole that spins when hit has been created-- the coins interactions corresponding to it are next. Prefabs and essential Game Objects have been imported to the main levels, including the Tutorial.