Checkpoint Prefab Updated
Checkpoints are now indicated by a mini version of the flagpole. Once a checkpoint is passed, players can no longer set that checkpoint as their respawn point if they touch a new checkpoint, then touch the old one. This is meant to future proof situations in Pachinko where the player falls down, passes an old checkpoint, then sets their spawn to the old checkpoint rather than the one further in the level.