GMs Deleted
There is a known bug with the Gallery where clearing the game makes the main menu inaccessible, and the only thing the player can do is quit. GMs across scenes have been deleted except for 2. It is unknown whether or not the TitleScreen's GM or the Hub's GM should be deleted to resolve this issue, but it should be solved as soon as humanly possible.