Phases 2 and 3 Completable
Phases 2 and 3 of the boss fight are now completable. A new copy of the button material has been created to differentiate hologram-spawning buttons from the level asset-activating buttons. Layouts for Phase 2 and 3 have been adjusted, including target positions. Players now reach the end cutscene as soon as the claw is hit 3 times. Some placeholder platform phases and targets have been established as a fail safe in case game logic does not work; they should not be deleted until the team is certain the scene transitions can work without them. A new tag called BossPlats has been created, but it may not be fully necessary; if not, it will be deleted in further updates.