I got distracted
Phase 2 of the Claw Machine boss battle has been reworked, as players were able to previously trivialize the phase by skipping the entire platform challenge with a well-timed wall jump on the end platform. This is no longer the case, as a gem barrier has been placed around the last platform that the player must deactivate by hitting another button.