1 bajillion things
all dialogue is in and functional! tutorial diag is now communicated through a super secret npc (pull to find out). claw diag is in and functional, but the npc himself isn't ready yet. all sound effects save for UI and missing animations are in. all current versions of music are in except for credits and technically main menu. updates to claw and coin pusher layouts, and some adjustments to colliders. checkpoints are easier to trigger and more frequent in coin pusher. possibly other things that i forgot ngl.